feat(update): improve how exceptions are rendered in a normalized way
