+++ app/Http/Controllers/UnauthenticatedController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\NormalizedException;
 use Closure;
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Http\Request;
@@ -15,10 +16,10 @@ class UnauthenticatedController extends Controller
         $this->middleware(function (Request $request, Closure $next) {
             // Validate if there is no user authenticated
             if (!is_null($request->user())) {
-                throw new TranslatableException(
+                throw new NormalizedException(
                     500,
                     'Authenticated user in unauthenticated route.',
-                    'api.ERROR.SOMETHING_WENT_WRONG',
+                    __('api.ERROR.SOMETHING_WENT_WRONG'),
                 );
             }
