+++ app/Http/Middleware/EnsureEmailIsVerified.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Middleware;
 
-use App\Exceptions\TranslatableException;
+use App\Exceptions\NormalizedException;
 use Closure;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -20,10 +20,10 @@ class EnsureEmailIsVerified
     public function handle(Request $request, Closure $next): Response
     {
         if (is_null($request->user()?->email_verified_at)) {
-            throw new TranslatableException(
+            throw new NormalizedException(
                 403,
                 'User email must be verified before accessing protected routes.',
-                'api.ERROR.EMAIL.NOT_VERIFIED'
+                __('api.ERROR.EMAIL.NOT_VERIFIED')
             );
         }
